requirements.txt
# Framework Django e REST
Django>=5.1.6
djangorestframework>=3.14.0